// src/services/types/tenant-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export type TenantResponse = {
	/**
	 * @description 테넌트 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 테넌트명
	 * @type string
	 */
	name: string;
	/**
	 * @description 생성자 아이디
	 * @type integer, int64
	 */
	creatorId: number;
	/**
	 * @description 생성 일시
	 * @type string, date-time
	 */
	createdAt: string;
};