// src/services/types/tenant-save-request.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export type TenantSaveRequest = {
	/**
	 * @description 테넌트 아이디 - null이면 등록, 아니면 수정
	 * @type integer | undefined, int64
	 */
	id?: number;
	/**
	 * @description 테넌트명
	 * @minLength 1
	 * @type string
	 */
	name: string;
};